% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{ct}
\alias{ct}
\title{Execute upon loading.}
\format{An object of class \code{V8} (inherits from \code{environment}) of length 8.}
\usage{
ct
}
\description{
Execute upon loading.
}
\keyword{datasets}

