% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomcolor.R
\name{randomColor}
\alias{randomColor}
\title{Get a pretty random color.}
\usage{
randomColor(count = 1, hue = c(" ", "random", "red", "orange", "yellow",
  "green", "blue", "purple", "pink", "monochrome"), luminosity = c(" ",
  "random", "light", "bright", "dark"))
}
\arguments{
\item{count}{number of colors (>= 1)}

\item{hue}{The hue of the colors to be selected.}

\item{luminosity}{The luminosity of the colors to be selected.}
}
\value{
A character vector of random color hexadecimal codes.
}
\description{
Get a pretty random color.
}
\examples{
randomColor()

randomColor(hue="pink")

randomColor(10, luminosity="light")
}
